requirements.txt
feedparser>=6.0.10
requests>=2.31.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.0
httpx>=0.25.0
